Update AI chat message styles
