+++ packages/twenty-front/src/modules/ai/utils/assistantMessageRenderItem.ts
@@ -1,0 +1,13 @@
+import { type ExtendedUIMessagePart } from 'twenty-shared/ai';
+
+import { type ThinkingStepPart } from '@/ai/utils/thinkingStepPart';
+
+export type AssistantMessageRenderItem =
+  | {
+      type: 'thinking-steps';
+      parts: ThinkingStepPart[];
+    }
+  | {
+      type: 'part';
+      part: ExtendedUIMessagePart;
+    };
