+++ packages/twenty-front/src/modules/ai/utils/thinkingStepPart.ts
@@ -1,0 +1,3 @@
+import { type ReasoningUIPart, type ToolUIPart } from 'ai';
+
+export type ThinkingStepPart = ReasoningUIPart | ToolUIPart;
