+++ packages/twenty-server/src/engine/core-modules/record-crud/zod-schemas/record-properties.zod-schema.ts
@@ -256,13 +256,11 @@ export const generateRecordPropertiesZodSchema = (
     }
 
     if (field.name === 'position') {
-      if (!forResponse) {
-        fieldSchema = z.union([
-          z.number(),
-          z.literal('first'),
-          z.literal('last'),
-        ]);
-      }
+      fieldSchema = z.union([
+        z.number(),
+        z.literal('first'),
+        z.literal('last'),
+      ]);
 
       fieldSchema = fieldSchema.describe(
         'Use "first" to insert at the top, "last" for the bottom, or a number for explicit ordering. Leave empty to place at the top (recommended).',
